Move variables unused/unrelated to store refs out of store module
